feat: Add generic main src code
This src comes from the examples. It is used to simply demonstrate that
KRPC builds and runs at a very basic level.

Technically, KSP with KRPC installed needs to be running on the same
machine for the code to actually work, but failing to connect to a
KSP instance still proves that the KRPC code is working correctly.