feat: Reorganize dockerfile to make use of caching
Files that change more often should be further down on the dockerfile.
The KRPC source code is expected to never change, and so can be placed
higher up on the dockerfile, so that it gets cached for later use.

Otherwise, every change to the FSW will also recompile KRPC, which is
not advisable.